Reorganise tests
Replace old tests with new (GitHub compatible) unit tests. Keep old tests around in `tests_old` for now so that we can check they still pass while we're migrating to the new tests.
