+++ .gitignore
@@ -15,11 +15,6 @@
 !tests/data/**/*.tls
 !tests/data/**/*.fits
 
-!tests_new/data/**/*.bsp
-!tests_new/data/**/*.tpc
-!tests_new/data/**/*.tls
-!tests_new/data/**/*.fits
-
 # Allowed paths
 !requirements.txt
 !.gitignore
