+++ tests/data/kernels/testing/nested/directory/hst.bsp
