+++ tests_old/test_spice_string_encoding.py
